+++ Kernel/System/GenericInterface/DebugLog.pm
@@ -726,6 +726,110 @@ sub LogSearch {
     return \@LogEntriesWithData;
 }
 
+=head2 LogCleanup()
+
+removes all log entries (including content) from a given time and before.
+
+returns 1 if successful or undef otherwise
+
+    my $Success = $DebugLogObject->LogCleanup(
+        CreatedAtOrBefore => '2011-12-31 23:59:59',
+    );
+
+=cut
+
+sub LogCleanup {
+    my ( $Self, %Param ) = @_;
+
+    if ( !$Param{CreatedAtOrBefore} ) {
+        $Kernel::OM->Get('Kernel::System::Log')->Log(
+            Priority => 'error',
+            Message  => "Need CreatedAtOrBefore",
+        );
+
+        return;
+    }
+
+    if ( $Param{CreatedAtOrBefore} !~ m{ \A \d{4} - \d{2} - \d{2} [ ] \d{2} : \d{2} : \d{2} \z }xms ) {
+
+        $Kernel::OM->Get('Kernel::System::Log')->Log(
+            Priority => 'error',
+            Message  => "CreatedAtOrBefore is not valid!",
+        );
+        return;
+    }
+
+    my $DateTimeObject = $Kernel::OM->Create('Kernel::System::DateTime');
+    my $Success = $DateTimeObject->Set( String => $Param{CreatedAtOrBefore} );
+    if ( !$Success ) {
+
+        $Kernel::OM->Get('Kernel::System::Log')->Log(
+            Priority => 'error',
+            Message  => "CreatedAtOrBefore is not valid!",
+        );
+        return;
+    }
+
+    my $DBObject = $Kernel::OM->Get('Kernel::System::DB');
+
+    # Get main debug log entries to delete
+    if (
+        !$DBObject->Prepare(
+            SQL  => 'SELECT id FROM gi_debugger_entry WHERE create_time <= ?',
+            Bind => [ \$Param{CreatedAtOrBefore} ],
+        )
+        )
+    {
+        $Kernel::OM->Get('Kernel::System::Log')->Log(
+            Priority => 'error',
+            Message  => 'Could not prepare db query!',
+        );
+        return;
+    }
+    my @LogEntryIDs;
+    while ( my @Row = $DBObject->FetchrowArray() ) {
+        push @LogEntryIDs, $Row[0];
+    }
+
+    return 1 if !@LogEntryIDs;
+
+    my $LogEntryIDsStr = join ',', @LogEntryIDs;
+
+    # Remove debug log entries contents.
+    if (
+        !$DBObject->Do(
+            SQL => "
+            DELETE FROM gi_debugger_entry_content
+            WHERE gi_debugger_entry_id in( $LogEntryIDsStr )",
+        )
+        )
+    {
+        $Kernel::OM->Get('Kernel::System::Log')->Log(
+            Priority => 'error',
+            Message  => 'Could not remove entries of communication chains in db!',
+        );
+        return;
+    }
+
+    # Remove debug log entries.
+    if (
+        !$DBObject->Do(
+            SQL => "
+            DELETE FROM gi_debugger_entry
+            WHERE id in( $LogEntryIDsStr )",
+        )
+        )
+    {
+        $Kernel::OM->Get('Kernel::System::Log')->Log(
+            Priority => 'error',
+            Message  => 'Could not remove communication chains in db!',
+        );
+        return;
+    }
+
+    return 1;
+}
+
 =begin Internal:
 
 =cut